:white_check_mark: Update ManagementAPI tests
