+++ test/2_management_api/correlations/get_active_correlations.js
@@ -11,7 +11,7 @@ describe('Management API: GetActiveCorrelations', () => {
 
   let defaultIdentity;
   let secondDefaultIdentity;
-  const processModelId = 'user_task_test';
+  const processModelId = 'test_management_api_usertask';
 
   before(async () => {
     testFixtureProvider = new TestFixtureProvider();
