+++ test/2_management_api/empty_activities/finish_empty_activity.js
@@ -4,7 +4,7 @@ const should = require('should');
 
 const {TestFixtureProvider, ProcessInstanceHandler} = require('../../../dist/commonjs/test_setup');
 
-describe(`Management API: FinishEmptyActivity`, () => {
+describe('Management API: FinishEmptyActivity', () => {
 
   let processInstanceHandler;
   let testFixtureProvider;
