+++ test/2_management_api/empty_activities/get_waiting_for_process_model_in_correlation.js
@@ -5,7 +5,7 @@ const uuid = require('node-uuid');
 
 const {TestFixtureProvider, ProcessInstanceHandler} = require('../../../dist/commonjs/test_setup');
 
-describe(`Management API: GetEmptyActivitiesForProcessModelInCorrelation`, () => {
+describe('Management API: GetEmptyActivitiesForProcessModelInCorrelation', () => {
 
   let eventAggregator;
   let processInstanceHandler;
@@ -119,7 +119,7 @@ describe(`Management API: GetEmptyActivitiesForProcessModelInCorrelation`, () =>
     before(async () => {
       // Create a number of ProcessInstances, so we can actually test pagination
       // We will have a grand total of 10 EmptyActivities after this.
-      for(let i = 0; i < 10; i++) {
+      for (let i = 0; i < 10; i++) {
         await processInstanceHandler.startProcessInstanceAndReturnResult(processModelId, correlationIdPaginationTest);
       }
       await processInstanceHandler.waitForProcessInstanceToReachSuspendedTask(correlationIdPaginationTest, processModelId, 10);
