+++ test/2_management_api/flow_node_instances/get_waiting_for_process_model_in_correlation.js
@@ -3,7 +3,7 @@
 const should = require('should');
 const uuid = require('node-uuid');
 
-const {ProcessInstanceHandler, TestFixtureProvider} = require('../../../dist/commonjs/test_setup');
+const {TestFixtureProvider, ProcessInstanceHandler} = require('../../../dist/commonjs/test_setup');
 
 describe('Management API: GetSuspendedTasksForProcessModelInCorrelation', () => {
